requirements.txt
<<<<<<< HEAD
flask
python-dotenv
google-genai
=======
flask
python-dotenv
google-genai
gunicorn


>>>>>>> 6490dfc476815c45cd8cfeca3c900e84e26d3b75
